+++ src/util/bip143.rs
@@ -194,7 +194,7 @@ mod tests {
     use consensus::encode::deserialize;
     use network::constants::Network;
     use util::address::Address;
-    use util::ecdsa::PublicKey;
+    use util::key::PublicKey;
     use hashes::hex::FromHex;
 
     use super::*;
